Add diagnostics using anyhow

+++ crates/parcel_core/src/types/source.rs
@@ -3,17 +3,17 @@ use std::path::PathBuf;
 use serde::Deserialize;
 use serde::Serialize;
 
-#[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq, Serialize)]
+#[derive(Clone, Debug, Default, Deserialize, Eq, Hash, PartialEq, Serialize)]
 pub struct Location {
-  pub line: u32,
-  pub column: u32,
+  pub line: usize,
+  pub column: usize,
 }
 
 /// Identifies a specific location in a source file
 ///
 /// Source locations start at 1:1.
 ///
-#[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq, Serialize)]
+#[derive(Clone, Debug, Default, Deserialize, Eq, Hash, PartialEq, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct SourceLocation {
   /// The file path associated with the source
